提现部分数据库double改int

+++ msip-bhu-business-impl/msip_bhu_business_ds/src/main/java/com/bhu/vas/business/ds/charging/facade/ChargingFacadeService.java
@@ -673,7 +673,7 @@ public class ChargingFacadeService {
 	}
 	
 	
-	public WithdrawCostInfo calculateWithdrawCost(int uid,String applyid,double cash){
+	public WithdrawCostInfo calculateWithdrawCost(int uid,String applyid,int cash){
 		return userWithdrawCostConfigsService.calculateWithdrawCost(uid,applyid,cash);
 	}
