+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/rpc/user/notify/IWalletNotifyCallback.java
@@ -3,5 +3,5 @@ package com.bhu.vas.api.rpc.user.notify;
 
 
 public interface IWalletNotifyCallback {
-	String notifyCashWithdrawOper(double cash);
+	String notifyCashWithdrawOper(int cash);
 }
